requirements.txt
datasets==2.21.0
mistralai
tqdm
soundfile
librosa
python-dotenv
